save stats as float64

+++ dl4d/datasets/ucr.py
@@ -88,7 +88,7 @@ class UCR(TimeseriesDataset):
         np.save(file=os.path.join(final_path, 'Y_val.npy'), arr=Y_val.astype(np.float32))
 
         shutil.rmtree(extracted_path)
-        self.save_stats(X_train.astype(np.float32))
+        self.save_stats(X_train.astype(np.float64))
         self.extract_features_from_npy()
